Code Clean Up

+++ app/src/main/java/com/scudderapps/e_tracker/Activity/UpdateActivity.java
@@ -1,15 +1,22 @@
 package com.scudderapps.e_tracker.Activity;
 
 import android.app.DatePickerDialog;
+import android.content.Context;
 import android.content.Intent;
+import android.graphics.Color;
 import android.os.Bundle;
+import android.text.Editable;
+import android.text.TextWatcher;
 import android.util.Patterns;
 import android.view.View;
+import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.DatePicker;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.textfield.TextInputEditText;
 import com.scudderapps.e_tracker.DATA.EmployeeData;
 import com.scudderapps.e_tracker.R;
@@ -21,12 +28,13 @@ import java.util.regex.Pattern;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class UpdateData extends AppCompatActivity {
+public class UpdateActivity extends AppCompatActivity {
 
     private static final Pattern NAME_PATTERN = Pattern.compile("^[A-Za-z\\s]{1,}[\\.]{0,1}[A-Za-z\\s]{0,}$");
     private static final Pattern PHONE_PATTERN = Pattern.compile("^[7-9][0-9]{9}$");
     private static final Pattern PASSWORD_PATTERN = Pattern.compile("^((?=.*[a-zA-Z]).{8,20})");
-    TextInputEditText searchEmpCode, empName, empPhone, empEmail, empPassword;
+
+    TextInputEditText searchEmpCode, empName, empPhone, empEmail, empPassword, empConfirmPassword;
     TextView empDob;
     Button searchEmpBtn, updateBtn, deleteBtn;
     String emp_dob_updated;
@@ -36,6 +44,7 @@ public class UpdateData extends AppCompatActivity {
     private Calendar c;
     private int month, day, year;
     private EmployeeData employeeData;
+    LinearLayout linearLayout;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -49,29 +58,67 @@ public class UpdateData extends AppCompatActivity {
         empEmail = findViewById(R.id.searchEmail);
         empPhone = findViewById(R.id.searchNumber);
         empPassword = findViewById(R.id.searchPassword);
+        empConfirmPassword = findViewById(R.id.searchConfirmPassword);
         empDob = findViewById(R.id.searchDob);
+        linearLayout = findViewById(R.id.updateLayout);
+        linearLayout.setVisibility(View.INVISIBLE);
         employeeData = new EmployeeData();
+        updateBtn.setEnabled(false);
+        deleteBtn.setEnabled(false);
 
+        empConfirmPassword.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void afterTextChanged(Editable mEdit) {
+                String cPass = mEdit.toString();
+                if (!cPass.isEmpty()){
+                    updateBtn.setEnabled(true);
+                    deleteBtn.setEnabled(true);
+                    updateBtn.setBackground(getResources().getDrawable(R.drawable.btn_bg));
+                    updateBtn.setTextColor(getResources().getColor(R.color.white));
+                } else {
+                    updateBtn.setEnabled(false);
+                    deleteBtn.setEnabled(false);
+                }
+            }
+
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+            }
+
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+            }
+        });
 
         searchEmpBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
 
                 emp_code = searchEmpCode.getText().toString();
+                InputMethodManager imm = (InputMethodManager) getSystemService(Context.INPUT_METHOD_SERVICE);
+                imm.hideSoftInputFromWindow(searchEmpBtn.getWindowToken(), InputMethodManager.RESULT_UNCHANGED_SHOWN);
                 searchList = MainActivity.employeeDatabase.employeeDAO().getSelectedEmployee(emp_code);
 
-                for (EmployeeData employeeData : searchList) {
-                    String emp_name = employeeData.getName();
-                    String emp_phone = employeeData.getPhone();
-                    String emp_email = employeeData.getEmail();
-                    String emp_pass = employeeData.getPassword();
-                    String emp_dob = employeeData.getDate();
+                if (!searchList.isEmpty()) {
+                    for (EmployeeData employeeData : searchList) {
+                        linearLayout.setVisibility(View.VISIBLE);
+                        String emp_name = employeeData.getName();
+                        String emp_phone = employeeData.getPhone();
+                        String emp_email = employeeData.getEmail();
+                        String emp_pass = employeeData.getPassword();
+                        String emp_dob = employeeData.getDate();
 
-                    empName.setText(emp_name);
-                    empPhone.setText(emp_phone);
-                    empEmail.setText(emp_email);
-                    empPassword.setText(emp_pass);
-                    empDob.setText(emp_dob);
+                        empName.setText(emp_name);
+                        empPhone.setText(emp_phone);
+                        empEmail.setText(emp_email);
+                        empPassword.setText(emp_pass);
+                        empDob.setText(emp_dob);
+                    }
+                } else {
+                    Snackbar snackbar = Snackbar.make(v, "No Records Found", Snackbar.LENGTH_LONG);
+                    View snackbarView = snackbar.getView();
+                    TextView textView = snackbarView.findViewById(R.id.snackbar_text);
+                    textView.setTextColor(Color.BLACK);
+                    snackbarView.setBackgroundColor(getResources().getColor(R.color.red));
+                    snackbar.show();
                 }
             }
         });
@@ -85,12 +132,12 @@ public class UpdateData extends AppCompatActivity {
         empDob.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                DatePickerDialog datePickerDialog = new DatePickerDialog(UpdateData.this, new DatePickerDialog.OnDateSetListener() {
+                DatePickerDialog datePickerDialog = new DatePickerDialog(UpdateActivity.this, new DatePickerDialog.OnDateSetListener() {
                     @Override
                     public void onDateSet(DatePicker view, int year, int monthOfYear, int dayOfMonth) {
                         c.set(year, monthOfYear, dayOfMonth);
-                        UpdateData.this.emp_dob_updated = dateFormat.format(c.getTime());
-                        empDob.setText(UpdateData.this.emp_dob_updated);
+                        UpdateActivity.this.emp_dob_updated = dateFormat.format(c.getTime());
+                        empDob.setText(UpdateActivity.this.emp_dob_updated);
                     }
                 }, year, month, day);
                 datePickerDialog.show();
@@ -119,11 +166,11 @@ public class UpdateData extends AppCompatActivity {
 
                     if (updatedDetails.size() != 0) {
                         MainActivity.employeeDatabase.employeeDAO().update(employeeData);
-                        Toast.makeText(UpdateData.this, "Details updated", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(UpdateActivity.this, "Details updated", Toast.LENGTH_SHORT).show();
                         back();
                     } else {
                         searchEmpCode.setError("Employee code not found");
-                        Toast.makeText(UpdateData.this, "No Employee Found", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(UpdateActivity.this, "No Employee Found", Toast.LENGTH_SHORT).show();
                     }
                 }
             }
@@ -132,9 +179,13 @@ public class UpdateData extends AppCompatActivity {
         deleteBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                MainActivity.employeeDatabase.employeeDAO().delete(emp_code);
-                Toast.makeText(UpdateData.this, "Employee Deleted", Toast.LENGTH_SHORT).show();
-                back();
+                if (validatePassword()) {
+                    MainActivity.employeeDatabase.employeeDAO().delete(emp_code);
+                    Toast.makeText(UpdateActivity.this, "Employee Deleted", Toast.LENGTH_SHORT).show();
+                    back();
+                } else {
+
+                }
             }
         });
     }
@@ -171,6 +222,7 @@ public class UpdateData extends AppCompatActivity {
 
     private boolean validatePassword() {
         String passwordInput = empPassword.getText().toString().trim();
+        String confirmPasswordInput = empConfirmPassword.getText().toString().trim();
 
         if (passwordInput.isEmpty()) {
             empPassword.setError("Field can't be empty");
@@ -180,6 +232,9 @@ public class UpdateData extends AppCompatActivity {
             return false;
         } else if (!PASSWORD_PATTERN.matcher(passwordInput).matches()) {
             empPassword.setError("Password too weak");
+            return false;
+        } else if (!passwordInput.equals(confirmPasswordInput)) {
+            empConfirmPassword.setError("Password do not match");
             return false;
         } else {
             empPassword.setError(null);
@@ -213,7 +268,7 @@ public class UpdateData extends AppCompatActivity {
     }
 
     public void back() {
-        Intent home = new Intent(UpdateData.this, MainActivity.class);
+        Intent home = new Intent(UpdateActivity.this, MainActivity.class);
         startActivity(home);
         finish();
     }
